Create a first real example, Bye Ya
The example is in closed voicings, but
it is a drill of the changes from the
A section of Bye-Ya to help me
get my feet wet.

I'm keeping it in the "learn" directory
because at this point I'm not doing any
of the things that I think I would do to
make the code scalable (such as extracting
individual voices so they can be recombined,
or more ambitiously using scheme).